Improve zip_mod.sh

+++ .scripts/zip_mod.sh
@@ -3,11 +3,10 @@
 
 
 ### Check commands
-has_errors=false
 if ! command -v git &> /dev/null; then
-	echo "Please install git https://git-scm.com/downloads"
-	has_errors=true
+	echo "Please install/use git https://git-scm.com/downloads"
 fi
+has_errors=false
 if ! command -v jq &> /dev/null; then
 	echo "Please install jq https://stedolan.github.io/jq/"
 	has_errors=true
@@ -21,9 +20,35 @@ if [ $has_errors = true ] ; then
 fi
 
 
-### REPOSITORY is current working directory
-REPOSITORY=`pwd`
-cd "$REPOSITORY/"
+### mod_folder is a mod directory with info.json
+init_dir=`pwd`
+
+
+### Find info.json
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+infojson_exists=false
+if [[ -s "$SCRIPT_DIR/info.json" ]]; then
+	infojson_exists=true
+	mod_folder=$SCRIPT_DIR
+else
+	cd $SCRIPT_DIR
+	cd ..
+	if [[ -s "$PWD/info.json" ]]; then
+		infojson_exists=true
+		mod_folder=$PWD
+	elif [[ -s "$init_dir/info.json" ]]; then
+		infojson_exists=true
+		mod_folder=$init_dir
+	fi
+fi
+
+
+if [ $infojson_exists = false ] ; then
+	echo "There's no info.json"
+	exit
+fi
+cd "$mod_folder/"
+echo "Target folder: ${mod_folder}"
 
 
 ### Get mod name and version from info.json
@@ -35,6 +60,7 @@ mod_ver=`cat info.json|jq -r .version`
 ### Prepare zip for Factorio native use and mod portal
 ### https://www.7-zip.org/download.html
 name="${mod_name}_${mod_ver}"
-git clean -xdf
-cd ..
-7z a -xr'!.*' "${REPOSITORY}/${name}.zip" "${REPOSITORY}"
+if command -v git &> /dev/null; then
+	git clean -xdf
+fi
+7z a -xr'!.*' "${mod_folder}/${name}.zip" "${mod_folder}"
